✨ H2 데이터베이스를 사용하는 ArticleRepository 추가
- ArticleEntity: jdbc 사용 시 setter 메서드가 불가피하다. 따라서 도메인 클래스 Article 가 무분별한 setter 메서드를 정의하고 있지 않도록 하기 위해 article 테이블과 1:1로 매핑할 수 있는 Entity 클래스를 추가한다.
- H2ArticleRepository 는 Repository 인터페이스의 구현체이며 @Primary 로 의존성 주입시 우선순위로 배정한다.

+++ src/main/java/com/kakao/cafe/service/ArticleService.java
@@ -9,5 +9,5 @@ public interface ArticleService {
 
     List<Article> searchAll();
     Article add(NewArticleParam newArticleParam);
-    Article search(int id);
+    Article search(long id);
 }
